add UserAuthenticatorHelper to mock CDX interaction w/o VCR
