+++ spec/requests/api/v0/post_tokens_spec.rb
@@ -1,36 +1,29 @@
 require 'rails_helper'
+include UserAuthenticatorHelper
 
 describe 'POST /api/v0/tokens' do
   it 'returns a 401 and no token when auth fails' do
-    VCR.use_cassette('auth_failure') do
-      user_credentials = { user_id: 'incorrect_user_id', password: 'incorrect_pw' }
-      token_params = { token: user_credentials }
+    mock_user_authenticator_fail
+    user_credentials = { user_id: 'incorrect_user_id', password: 'incorrect_pw' }
+    token_params = { token: user_credentials }
 
-      post '/api/v0/tokens',
-        token_params.to_json,
-        set_headers
+    post '/api/v0/tokens', token_params.to_json, set_headers
 
-      expect(response.status).to eq(401)
-      expect(parsed_response['token']).to be_nil
-      expect(parsed_response['errors']).to eq(
-        'Unable to authenticate user - The user account could not be located. Could not perform central authentication for user <change_me> in domain <default>'
-      )
-    end
+    expect(response.status).to eq(401)
+    expect(parsed_response['token']).to be_nil
+    expect(parsed_response['errors']).to eq('Bad user_id or password')
   end
 
   it 'returns token on success' do
-    VCR.use_cassette('auth_success') do
-      user_credentials = { user_id: 'correct_username', password: 'correct_pw' }
-      token_params = { token: user_credentials }
+    mock_user_authenticator_pass
+    user_credentials = { user_id: 'correct_username', password: 'correct_pw' }
+    token_params = { token: user_credentials }
 
-      post '/api/v0/tokens',
-        token_params.to_json,
-        set_headers
+    post '/api/v0/tokens', token_params.to_json, set_headers
 
-      expect(response.status).to eq(200)
-      expect(parsed_response['token']).not_to be_nil
-      expect(Redis.new.get(parsed_response['token'])).to_not be_nil
-     end
+    expect(response.status).to eq(200)
+    expect(parsed_response['token']).not_to be_nil
+    expect(UserSession.new(parsed_response['token']).user).to be_a User
   end
 
   private
