+++ spec/services/user_authenticator_spec.rb
@@ -1,28 +1,25 @@
 require 'rails_helper'
+include UserAuthenticatorHelper
 
 describe UserAuthenticator do
-  before(:each) do
-    user = create(:user)
-    session = UserSession.create(user, { token: SecureRandom.hex })
-    allow_any_instance_of(UserAuthenticator).to receive(:authenticate).and_return(session)
-    allow_any_instance_of(UserAuthenticator).to receive(:cdx_response).and_return(session.cdx_auth_response)
-    allow_any_instance_of(UserAuthenticator).to receive(:error_message).and_return('Bad user_id or password')
-  end
-
   describe 'workflow' do
     it '#authenticate' do
+      mock_user_authenticator_pass
       authenticator = UserAuthenticator.new(user_id: 'user', password: 'pass')
       expect(authenticator.authenticate).to be_a UserSession
     end
 
     it '#cdx_response' do
+      mock_user_authenticator_pass
       authenticator = UserAuthenticator.new(user_id: 'user', password: 'pass')
       authenticator.authenticate
       expect(authenticator.cdx_response[:token]).to_not be_nil
     end
 
     it '#error_message' do
+      mock_user_authenticator_fail
       authenticator = UserAuthenticator.new(user_id: 'nope', password: 'never')
+      expect(authenticator.authenticate).to be_nil
       expect(authenticator.error_message).to eq 'Bad user_id or password'
     end
