--- spec/support/request_helper.rb
@@ -1,5 +1,0 @@
-module RequestHelper
-  def login_as(user, cdx_response = nil)
-    UserSession.create(user, cdx_response)
-  end
-end
